Operate on inputs and display outputs
Inputs are taken and outputs are given but the final issue is
with updating the operator. I have two binaries: one to say if there
have been any inputs yet and another to say if we should carry on
with the calculation as if we already have an input.

When we have an input and keep going with calculations, we never reach
the code that updates the operation. I am thinking of solutions.